added funcs, step one of merge
